refactor(mod): simplify main by adding cli and astarte modules
